Massive code refactoring
